requirements.txt
asgiref==3.11.1
Django==5.2.13
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.12
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
tzdata==2026.2
